+++ src/service/payment/movieTicket.ts
@@ -166,8 +166,7 @@ export function authorize(params: {
             if (movieTheater.paymentAccepted === undefined) {
                 throw new factory.errors.Argument('transaction', 'Movie Ticket payment not accepted');
             }
-            const movieTicketPaymentAccepted = <factory.seller.IMovieTicketPaymentAccepted>
-                movieTheater.paymentAccepted.find((a) => a.paymentMethodType === paymentMethodType);
+            const movieTicketPaymentAccepted = movieTheater.paymentAccepted.find((a) => a.paymentMethodType === paymentMethodType);
             if (movieTicketPaymentAccepted === undefined) {
                 throw new factory.errors.Argument('transaction', 'Movie Ticket payment not accepted');
             }
@@ -328,8 +327,7 @@ export function checkMovieTicket(
             if (movieTheater.paymentAccepted === undefined) {
                 throw new factory.errors.Argument('transactionId', 'Movie Ticket payment not accepted');
             }
-            const movieTicketPaymentAccepted = <factory.seller.IMovieTicketPaymentAccepted>
-                movieTheater.paymentAccepted.find((a) => a.paymentMethodType === paymentMethodType);
+            const movieTicketPaymentAccepted = movieTheater.paymentAccepted.find((a) => a.paymentMethodType === paymentMethodType);
             if (movieTicketPaymentAccepted === undefined) {
                 throw new factory.errors.Argument('transactionId', 'Movie Ticket payment not accepted');
             }